Update README.md
In these exercises, we'll stimulate your curiosity and train your knowledge of the digital world. Think of these exercises as training!
If you want to become an explorer like the crew of SAYNA, you've got to go through this. To give you an example, the astronauts of your time don't go into space overnight. They undergo intense training, both physically and mentally, to cope with the hostile environment of space.

The digital world is just as incredible and hostile as space. You have a multitude of things to discover, but if you're unprepared, you run the risk of running into a number of pitfalls. We invite you to sharpen your senses and explore this vast universe.

Play Hard Work Hard